Fix reading mpd binary responses
The `fread` call never returns more then ~16k bytes, at least on my
system. Even `stream_set_read_buffer` does not fix that, so there must
be some other buffer which limits the reads. To work around this, check
the amount of data we got against what was requested. If we get less
data then we requested, repeat the read until we got everything we want.

The `binarylimit` command was added in mpd v0.22.4, so only use it when
available. It is not necessary, it will just speed things up a bit.